Resolved issue w/ tran_outer by moving it from tran_tools.py to tools.py
