+++ grama/tran_tools.py
@@ -5,8 +5,8 @@ __all__ = [
     "tf_bootstrap",
     "tran_copula_corr",
     "tf_copula_corr",
-    "tran_outer",
-    "tf_outer",
+    # "tran_outer",
+    # "tf_outer",
     "tran_kfolds",
     "tf_kfolds",
     "tran_md",
@@ -287,46 +287,46 @@ def tran_bootstrap(
 tf_bootstrap = add_pipe(tran_bootstrap)
 
 
-## DataFrame outer product
-# --------------------------------------------------
-@curry
-def tran_outer(df, df_outer):
-    r"""Outer merge
-
-    Perform an outer-merge on two dataframes.
-
-    Args:
-        df (DataFrame): Data to merge
-        df_outer (DataFrame): Data to merge; outer
-
-    Returns:
-        DataFrame: Merged data
-
-    Examples:
-        >>> import grama as gr
-        >>> import pandas as pd
-        >>> df = pd.DataFrame(dict(x=[1,2]))
-        >>> df_outer = pd.DataFrame(dict(y=[3,4]))
-        >>> df_res = gr.tran_outer(df, df_outer)
-        >>> df_res
-        >>>    x  y
-        >>> 0  1  3
-        >>> 1  2  3
-        >>> 2  1  4
-        >>> 3  2  4
-
-    """
-    n_rows = df.shape[0]
-    list_df = []
-
-    for ind in range(df_outer.shape[0]):
-        df_rep = concat([df_outer.iloc[[ind]]] * n_rows, ignore_index=True)
-        list_df.append(concat((df, df_rep), axis=1))
-
-    return concat(list_df, ignore_index=True)
-
-
-tf_outer = add_pipe(tran_outer)
+# ## DataFrame outer product
+# # --------------------------------------------------
+# @curry
+# def tran_outer(df, df_outer):
+#     r"""Outer merge
+#
+#     Perform an outer-merge on two dataframes.
+#
+#     Args:
+#         df (DataFrame): Data to merge
+#         df_outer (DataFrame): Data to merge; outer
+#
+#     Returns:
+#         DataFrame: Merged data
+#
+#     Examples:
+#         >>> import grama as gr
+#         >>> import pandas as pd
+#         >>> df = pd.DataFrame(dict(x=[1,2]))
+#         >>> df_outer = pd.DataFrame(dict(y=[3,4]))
+#         >>> df_res = gr.tran_outer(df, df_outer)
+#         >>> df_res
+#         >>>    x  y
+#         >>> 0  1  3
+#         >>> 1  2  3
+#         >>> 2  1  4
+#         >>> 3  2  4
+#
+#     """
+#     n_rows = df.shape[0]
+#     list_df = []
+#
+#     for ind in range(df_outer.shape[0]):
+#         df_rep = concat([df_outer.iloc[[ind]]] * n_rows, ignore_index=True)
+#         list_df.append(concat((df, df_rep), axis=1))
+#
+#     return concat(list_df, ignore_index=True)
+#
+#
+# tf_outer = add_pipe(tran_outer)
 
 
 ## Assess subspace angles
